services/identity_service/pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Identity and auth service for fintech practice stack"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "pydantic",
    "pydantic-settings",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "redis",
    "loguru",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-sdk",
    "prometheus-client"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
